backend/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
argcomplete==3.1.4
asgiref==3.7.2
astroid==3.2.2
attrs==23.2.0
Automat==22.10.0
autopep8==2.3.1
Babel==2.10.3
bcrypt==3.2.2
bleach==6.1.0
blinker==1.7.0
certifi==2023.11.17
channels==4.1.0
channels-redis==4.2.0
chardet==5.2.0
click==8.1.6
cloud-init==24.1.3
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
constantly==23.10.4
cryptography==41.0.7
cssselect==1.2.0
dbus-python==1.3.2
decorator==5.1.1
dicttoxml==1.7.16
dill==0.3.8
distro==1.9.0
distro-info==1.7+build1
Django==4.2.11
django-cors-headers==4.3.1
django-extensions==3.2.3
django-js-asset==2.2.0
django-mptt==0.16.0
django-sslserver==0.22
djangorestframework==3.15.1
docutils==0.21.2
frozenlist==1.4.1
future==1.0.0
GDAL==3.8.4
geocoder==1.38.1
geographiclib==2.0
geoip2==4.8.0
geopy==2.4.1
h11==0.14.0
httplib2==0.20.4
hyperlink==21.0.0
idna==3.6
importlib_metadata==8.0.0
incremental==22.10.0
ip2geotools==0.1.6
IP2Location==8.10.2
isort==5.13.2
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jeepney==0.8.0
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
keyring==25.2.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
lazy-object-proxy==1.10.0
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
maxminddb==2.6.2
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.3.0
msgpack==1.0.8
multidict==6.0.5
mysqlclient==2.2.4
netifaces==0.11.0
nh3==0.2.17
numpy==1.26.4
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.0
pillow==10.3.0
pip-review==1.3.0
pipx==1.4.3
pkginfo==1.10.0
platformdirs==4.2.0
psutil==5.9.8
psycopg2-binary==2.9.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.12.0
pycurl==7.45.3
Pygments==2.17.2
PyGObject==3.48.2
PyHamcrest==2.1.0
PyJWT==2.7.0
pylint==3.2.4
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyquery==2.0.0
pyrsistent==0.20.0
pyserial==3.5
PySocks==1.7.1
python-apt==2.7.7+ubuntu1
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
ratelim==0.1.6
readme_renderer==43.0
redis==5.0.7
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
SecretStorage==3.3.3
selenium==4.22.0
service-identity==24.1.0
setuptools==68.1.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.4.4
systemd-python==235
tomlkit==0.12.5
tqdm==4.66.4
trio==0.25.1
trio-websocket==0.11.1
twine==5.1.1
Twisted==24.3.0
typed-ast==1.5.5
typing==3.7.4.3
typing_extensions==4.10.0
ubuntu-pro-client==8001
unattended-upgrades==0.1
urllib3==2.0.7
userpath==1.9.1
wadllib==1.3.6
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
wheel==0.42.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.19.2
zope.interface==6.1
